Add files via upload
Managed to add a media file so that I can show images when creating a recipe. Still havent been  able to get the password recovery working but trying. While waiting managed to get the documentation done to which I can always add to if needed. Terminal is causing too many issue especially since the video is too big and the gitignore isnt ignoring it